moving negatePolynomialTerm to Negative module instead of the PolynomialTermOperations module
