+++ lib/expression/step-solver/Negative.js
@@ -3,7 +3,6 @@
 const NodeCreator = require('./NodeCreator');
 const NodeType = require('./NodeType');
 const PolynomialTermNode = require('./PolynomialTermNode');
-const PolynomialTermOperations = require('./PolynomialTermOperations');
 
 const Negative = {};
 
@@ -44,7 +43,7 @@ Negative.negate = function(node, naive=false) {
     return node;
   }
   else if (PolynomialTermNode.isPolynomialTerm(node)) {
-    return PolynomialTermOperations.negatePolynomialTerm(node, naive);
+    return Negative.negatePolynomialTerm(node, naive);
   }
   else if (!naive) {
     if (NodeType.isUnaryMinus(node)) {
@@ -56,6 +55,45 @@ Negative.negate = function(node, naive=false) {
   }
 
   return NodeCreator.unaryMinus(node);
+}
+
+// Multiplies a polynomial term by -1 and returns the new node
+// If naive is true, then we just add an extra unary minus to the expression
+// otherwise, we do the actual negation
+// E.g.
+//    not naive: -3x -> 3x, x -> -x
+//    naive: -3x -> --3x, x -> -x
+Negative.negatePolynomialTerm = function(node, naive=false) {
+  if(!PolynomialTermNode.isPolynomialTerm(node)) {
+    throw Error('node is not a polynomial term');
+  }
+  const polyNode = new PolynomialTermNode(node);
+
+  let newCoeff;
+  if (!polyNode.hasCoeff()) {
+    newCoeff = NodeCreator.constant(-1);
+  }
+  else {
+    const oldCoeff = polyNode.getCoeffNode();
+    if (oldCoeff.value === '-1') {
+      newCoeff = null;
+    }
+    else if (polyNode.hasFractionCoeff()) {
+      let numerator = oldCoeff.args[0];
+      numerator = Negative.negate(numerator, naive);
+
+      const denominator = oldCoeff.args[1];
+      newCoeff = NodeCreator.operator('/', [numerator, denominator]);
+    }
+    else {
+      newCoeff = Negative.negate(oldCoeff, naive);
+      if (newCoeff.value === '1') {
+        newCoeff = null;
+      }
+    }
+  }
+  return NodeCreator.polynomialTerm(
+    polyNode.getSymbolNode(), polyNode.getExponentNode(), newCoeff);
 }
 
 module.exports = Negative;
