+++ lib/expression/step-solver/PolynomialTermOperations.js
@@ -65,58 +65,6 @@ PolynomialTermOperations.multiplyConstantAndPolynomialTerm = function(node) {
   }
 }
 
-// Multiplies a polynomial term by -1 and returns the new node
-// If naive is true, then we just add an extra unary minus to the expression
-// otherwise, we do the actual negation
-// E.g.
-//    not naive: -3x -> 3x, x -> -x
-//    naive: -3x -> --3x, x -> -x
-PolynomialTermOperations.negatePolynomialTerm = function(node, naive=false) {
-  if(!PolynomialTermNode.isPolynomialTerm(node)) {
-    throw Error('node is not a polynomial term');
-  }
-  const polyNode = new PolynomialTermNode(node);
-
-  let newCoeff;
-  if (!polyNode.hasCoeff()) {
-    newCoeff = NodeCreator.constant(-1);
-  }
-  else {
-    const oldCoeff = polyNode.getCoeffNode();
-    if (oldCoeff.value === '-1') {
-      newCoeff = null;
-    }
-    else if (polyNode.hasFractionCoeff()) {
-      let numerator = oldCoeff.args[0];
-      if (naive) {
-        numerator = NodeCreator.unaryMinus(numerator);
-      }
-      else {
-        numerator = NodeCreator.constant(0 - parseFloat(numerator.value));
-      }
-
-      const denominator = oldCoeff.args[1];
-      newCoeff = NodeCreator.operator('/', [numerator, denominator]);
-    }
-    else {
-      if (naive) {
-        newCoeff = NodeCreator.unaryMinus(oldCoeff);
-      }
-      else {
-        const newCoeffValue = 0 - parseFloat(oldCoeff.value);
-        if (newCoeffValue === 1) {
-          newCoeff = null;
-        }
-        else {
-          newCoeff = NodeCreator.constant(newCoeffValue);
-        }
-      }
-    }
-  }
-  return NodeCreator.polynomialTerm(
-    polyNode.getSymbolNode(), polyNode.getExponentNode(), newCoeff);
-}
-
 // Simplifies a polynomial term with a fraction as its coefficients.
 // e.g. 2x/4 --> x/2    10x/5 --> 2x
 // Also simplified negative signs
